BJS2-77776: fix all comment in PR
