+++ src/main/java/school/faang/user_service/controller/handler/RestControllerExceptionHandler.java
@@ -5,16 +5,14 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.context.request.WebRequest;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 import school.faang.user_service.exception.DataValidationException;
 
 @Slf4j
-@RestControllerAdvice(basePackages = "school.faang.user_service.controller")
-public class RestControllerExceptionHandler extends ResponseEntityExceptionHandler {
+@RestControllerAdvice
+public class RestControllerExceptionHandler {
 
     @ExceptionHandler(DataValidationException.class)
-    public final ResponseEntity<Object> handleDataValidationException(DataValidationException exception, WebRequest request) {
+    public final ResponseEntity<Object> handleDataValidationException(DataValidationException exception) {
         log.warn("Validation error: {}", exception.getMessage());
         return ResponseEntity
                 .status(HttpStatus.BAD_REQUEST)
