+++ src/main/java/school/faang/user_service/service/SubscriptionService.java
@@ -1,65 +1,21 @@
 package school.faang.user_service.service;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 import school.faang.user_service.dto.UserDto;
 import school.faang.user_service.dto.UserFilterDto;
-import school.faang.user_service.entity.User;
-import school.faang.user_service.exception.DataValidationException;
-import school.faang.user_service.filter.UserFilter;
-import school.faang.user_service.mapper.UserMapper;
-import school.faang.user_service.repository.SubscriptionRepository;
 
 import java.util.List;
-import java.util.stream.Stream;
 
-@Service
-@RequiredArgsConstructor
-public class SubscriptionService {
+public interface SubscriptionService {
 
-    private final SubscriptionRepository subscriptionRepository;
-    private final List<UserFilter> userFilters;
-    private final UserMapper userMapper;
+    void followUser(long followerId, long followeeId);
 
-    public void followUser(long followerId, long followeeId) {
-        boolean exists = subscriptionRepository.existsByFollowerIdAndFolloweeId(followerId, followeeId);
-        if (exists) {
-            throw new DataValidationException("You are already following this user");
-        }
-        subscriptionRepository.followUser(followerId, followeeId);
-    }
+    void unfollowUser(long followerId, long followeeId);
 
-    public void unfollowUser(long followerId, long followeeId) {
-        subscriptionRepository.unfollowUser(followerId, followeeId);
-    }
+    List<UserDto> getFollowers(long followeeId, UserFilterDto userFilterDto);
 
-    public List<UserDto> getFollowers(long followeeId, UserFilterDto userFilterDto) {
-        Stream<User> followers = subscriptionRepository.findByFolloweeId(followeeId);
-        return filterUsers(followers, userFilterDto)
-                .map(userMapper::toUserDTO)
-                .toList();
-    }
+    int getFollowersCount(long followeeId);
 
-    private Stream<User> filterUsers(Stream<User> users, UserFilterDto userFilterDto) {
-        return userFilters.stream()
-                .filter(filter -> filter.isApplicable(userFilterDto))
-                .reduce(users,
-                        (currentStream, filter) -> filter.apply(currentStream, userFilterDto),
-                        Stream::concat);
-    }
+    List<UserDto> getFollowing(long followerId, UserFilterDto userFilterDto);
 
-    public int getFollowersCount(long followeeId) {
-        return subscriptionRepository.findFollowersAmountByFolloweeId(followeeId);
-    }
-
-    public List<UserDto> getFollowing(long followerId, UserFilterDto userFilterDto) {
-        Stream<User> followees = subscriptionRepository.findByFollowerId(followerId);
-        return filterUsers(followees, userFilterDto)
-                .map(userMapper::toUserDTO)
-                .toList();
-    }
-
-    public int getFollowingCount(long followerId) {
-        return subscriptionRepository.findFolloweesAmountByFollowerId(followerId);
-    }
+    int getFollowingCount(long followerId);
 }
