+++ src/test/java/school/faang/user_service/service/SubscriptionServiceImplTest.java
@@ -18,7 +18,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.when;
 
 @SpringBootTest
-class SubscriptionServiceTest {
+class SubscriptionServiceImplTest {
 
     @MockBean
     private SubscriptionRepository subscriptionRepository;
